Extract initialization of integration test
